Refactor malingering FragmentingUdpLink to ephemeral FragmentingUdpSocket.
