+++ src/commonMain/kotlin/baaahs/di/Modules.kt
@@ -5,7 +5,6 @@ import baaahs.dmx.Dmx
 import baaahs.gl.render.RenderManager
 import baaahs.io.Fs
 import baaahs.model.Model
-import baaahs.net.FragmentingUdpLink
 import baaahs.net.Network
 import baaahs.plugin.PluginContext
 import baaahs.plugin.Plugins
@@ -14,7 +13,10 @@ import baaahs.plugin.beatlink.BeatSource
 import baaahs.proto.Ports
 import baaahs.sim.FakeDmxUniverse
 import baaahs.util.Clock
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.SupervisorJob
 import org.koin.core.module.Module
 import org.koin.core.qualifier.named
 import org.koin.core.scope.Scope
@@ -43,7 +45,7 @@ interface PinkyModule : KModule {
     val Scope.firmwareDir: Fs.File
     val Scope.firmwareDaddy: FirmwareDaddy
     val Scope.pinkyMainDispatcher: CoroutineDispatcher
-    val Scope.pinkyLink: Network.Link get() = FragmentingUdpLink(get<Network>().link("pinky"))
+    val Scope.pinkyLink: Network.Link get() = get<Network>().link("pinky")
     val Scope.dmxUniverse: Dmx.Universe
     val Scope.renderManager: RenderManager
